+++ vectorstorefile.go
@@ -94,10 +94,10 @@ func (r *VectorStoreFileService) UploadAndPoll(ctx context.Context, vectorStoreI
 }
 
 // Retrieves a vector store file.
-func (r *VectorStoreFileService) Get(ctx context.Context, vectorStoreID string, fileID string, opts ...option.RequestOption) (res *VectorStoreFile, err error) {
+func (r *VectorStoreFileService) Get(ctx context.Context, fileID string, query VectorStoreFileGetParams, opts ...option.RequestOption) (res *VectorStoreFile, err error) {
 	opts = append(r.Options[:], opts...)
 	opts = append([]option.RequestOption{option.WithHeader("OpenAI-Beta", "assistants=v2")}, opts...)
-	if vectorStoreID == "" {
+	if query.VectorStoreID == "" {
 		err = errors.New("missing required vector_store_id parameter")
 		return
 	}
@@ -105,16 +105,16 @@ func (r *VectorStoreFileService) Get(ctx context.Context, vectorStoreID string, 
 		err = errors.New("missing required file_id parameter")
 		return
 	}
-	path := fmt.Sprintf("vector_stores/%s/files/%s", vectorStoreID, fileID)
+	path := fmt.Sprintf("vector_stores/%s/files/%s", query.VectorStoreID, fileID)
 	err = requestconfig.ExecuteNewRequest(ctx, http.MethodGet, path, nil, &res, opts...)
 	return
 }
 
 // Update attributes on a vector store file.
-func (r *VectorStoreFileService) Update(ctx context.Context, vectorStoreID string, fileID string, body VectorStoreFileUpdateParams, opts ...option.RequestOption) (res *VectorStoreFile, err error) {
+func (r *VectorStoreFileService) Update(ctx context.Context, fileID string, params VectorStoreFileUpdateParams, opts ...option.RequestOption) (res *VectorStoreFile, err error) {
 	opts = append(r.Options[:], opts...)
 	opts = append([]option.RequestOption{option.WithHeader("OpenAI-Beta", "assistants=v2")}, opts...)
-	if vectorStoreID == "" {
+	if params.VectorStoreID == "" {
 		err = errors.New("missing required vector_store_id parameter")
 		return
 	}
@@ -122,8 +122,8 @@ func (r *VectorStoreFileService) Update(ctx context.Context, vectorStoreID strin
 		err = errors.New("missing required file_id parameter")
 		return
 	}
-	path := fmt.Sprintf("vector_stores/%s/files/%s", vectorStoreID, fileID)
-	err = requestconfig.ExecuteNewRequest(ctx, http.MethodPost, path, body, &res, opts...)
+	path := fmt.Sprintf("vector_stores/%s/files/%s", params.VectorStoreID, fileID)
+	err = requestconfig.ExecuteNewRequest(ctx, http.MethodPost, path, params, &res, opts...)
 	return
 }
 
@@ -158,10 +158,10 @@ func (r *VectorStoreFileService) ListAutoPaging(ctx context.Context, vectorStore
 // the file itself will not be deleted. To delete the file, use the
 // [delete file](https://platform.openai.com/docs/api-reference/files/delete)
 // endpoint.
-func (r *VectorStoreFileService) Delete(ctx context.Context, vectorStoreID string, fileID string, opts ...option.RequestOption) (res *VectorStoreFileDeleted, err error) {
+func (r *VectorStoreFileService) Delete(ctx context.Context, fileID string, body VectorStoreFileDeleteParams, opts ...option.RequestOption) (res *VectorStoreFileDeleted, err error) {
 	opts = append(r.Options[:], opts...)
 	opts = append([]option.RequestOption{option.WithHeader("OpenAI-Beta", "assistants=v2")}, opts...)
-	if vectorStoreID == "" {
+	if body.VectorStoreID == "" {
 		err = errors.New("missing required vector_store_id parameter")
 		return
 	}
@@ -169,17 +169,17 @@ func (r *VectorStoreFileService) Delete(ctx context.Context, vectorStoreID strin
 		err = errors.New("missing required file_id parameter")
 		return
 	}
-	path := fmt.Sprintf("vector_stores/%s/files/%s", vectorStoreID, fileID)
+	path := fmt.Sprintf("vector_stores/%s/files/%s", body.VectorStoreID, fileID)
 	err = requestconfig.ExecuteNewRequest(ctx, http.MethodDelete, path, nil, &res, opts...)
 	return
 }
 
 // Retrieve the parsed contents of a vector store file.
-func (r *VectorStoreFileService) Content(ctx context.Context, vectorStoreID string, fileID string, opts ...option.RequestOption) (res *pagination.Page[VectorStoreFileContentResponse], err error) {
+func (r *VectorStoreFileService) Content(ctx context.Context, fileID string, query VectorStoreFileContentParams, opts ...option.RequestOption) (res *pagination.Page[VectorStoreFileContentResponse], err error) {
 	var raw *http.Response
 	opts = append(r.Options[:], opts...)
 	opts = append([]option.RequestOption{option.WithHeader("OpenAI-Beta", "assistants=v2"), option.WithResponseInto(&raw)}, opts...)
-	if vectorStoreID == "" {
+	if query.VectorStoreID == "" {
 		err = errors.New("missing required vector_store_id parameter")
 		return
 	}
@@ -187,7 +187,7 @@ func (r *VectorStoreFileService) Content(ctx context.Context, vectorStoreID stri
 		err = errors.New("missing required file_id parameter")
 		return
 	}
-	path := fmt.Sprintf("vector_stores/%s/files/%s/content", vectorStoreID, fileID)
+	path := fmt.Sprintf("vector_stores/%s/files/%s/content", query.VectorStoreID, fileID)
 	cfg, err := requestconfig.NewRequestConfig(ctx, http.MethodGet, path, nil, &res, opts...)
 	if err != nil {
 		return nil, err
@@ -201,8 +201,8 @@ func (r *VectorStoreFileService) Content(ctx context.Context, vectorStoreID stri
 }
 
 // Retrieve the parsed contents of a vector store file.
-func (r *VectorStoreFileService) ContentAutoPaging(ctx context.Context, vectorStoreID string, fileID string, opts ...option.RequestOption) *pagination.PageAutoPager[VectorStoreFileContentResponse] {
-	return pagination.NewPageAutoPager(r.Content(ctx, vectorStoreID, fileID, opts...))
+func (r *VectorStoreFileService) ContentAutoPaging(ctx context.Context, fileID string, query VectorStoreFileContentParams, opts ...option.RequestOption) *pagination.PageAutoPager[VectorStoreFileContentResponse] {
+	return pagination.NewPageAutoPager(r.Content(ctx, fileID, query, opts...))
 }
 
 // A list of files attached to a vector store.
@@ -433,13 +433,19 @@ func (u *VectorStoreFileNewParamsAttributeUnion) asAny() any {
 	return nil
 }
 
+type VectorStoreFileGetParams struct {
+	VectorStoreID string `path:"vector_store_id,required" json:"-"`
+	paramObj
+}
+
 type VectorStoreFileUpdateParams struct {
 	// Set of 16 key-value pairs that can be attached to an object. This can be useful
 	// for storing additional information about the object in a structured format, and
 	// querying for objects via API or the dashboard. Keys are strings with a maximum
 	// length of 64 characters. Values are strings with a maximum length of 512
 	// characters, booleans, or numbers.
-	Attributes map[string]VectorStoreFileUpdateParamsAttributeUnion `json:"attributes,omitzero,required"`
+	Attributes    map[string]VectorStoreFileUpdateParamsAttributeUnion `json:"attributes,omitzero,required"`
+	VectorStoreID string                                               `path:"vector_store_id,required" json:"-"`
 	paramObj
 }
 
@@ -532,3 +538,13 @@ const (
 	VectorStoreFileListParamsOrderAsc  VectorStoreFileListParamsOrder = "asc"
 	VectorStoreFileListParamsOrderDesc VectorStoreFileListParamsOrder = "desc"
 )
+
+type VectorStoreFileDeleteParams struct {
+	VectorStoreID string `path:"vector_store_id,required" json:"-"`
+	paramObj
+}
+
+type VectorStoreFileContentParams struct {
+	VectorStoreID string `path:"vector_store_id,required" json:"-"`
+	paramObj
+}